# Docs/api.yaml
openapi: 3.1.0
info:
  title: Swagger SpaceX - OpenAPI 3.0
  description: |-
    This is a description of SpaceX api on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io).

  version: 0.1.0
...
servers:
  - url: http://127.0.0.1:8080/api/v1
paths:
...
  /addToBasket:
    post:
      summary: Add a travel to the basket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                basketId:
                  type: "String"
                  example: "bf8e97ea-c46a-46d8-9c88-3b5289d0df25"
                travelId:
                  type: "integer"
                  example: 1
      responses:
        '200':
          description: Travel added to basket
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TravelItem'
  /removeFromBasket:
    post:
      summary: Remove a travel from the basket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                basketId:
                  type: "String"
                  example: "bf8e97ea-c46a-46d8-9c88-3b5289d0df25"
                travelId:
                  type: "integer"
                  example: 1
      responses:
        '200':
          description: Travel removed from basket
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TravelItem'
...
components:
  schemas:
    TravelItem:
      type: object
      properties:
        id:
          type: integer
          example: 125
        depart:
          type: string
          example: MRC
        arrive:
          type: string
          example: TRR
        dateDepart:
          type: string
          format: date
          example: 12/12/2023
        dateArrive:
          type: string
          format: date
          example: 06/06/2024
        prix:
          type: integer
          example: 45212
...